#127: re-factor resource utils between common/unit/integration tests
