+++ src/test/common/java/com/commercetools/testUtil/ResourcesUtil.java
@@ -1,9 +1,12 @@
 package com.commercetools.testUtil;
 
-public class ResourcesUtil {
+public final class ResourcesUtil {
     public static String MOCK_ROOT_DIR = "mockData/";
 
     public static String resolveMockDataResource(String mockDataRelativePath) {
         return MOCK_ROOT_DIR + mockDataRelativePath;
+    }
+
+    private ResourcesUtil() {
     }
 }
