+++ src/test/integration/java/com/commercetools/service/ctp/impl/OrderServiceImplIT.java
@@ -2,7 +2,6 @@ package com.commercetools.service.ctp.impl;
 
 import com.commercetools.Application;
 import com.commercetools.service.ctp.OrderService;
-import com.commercetools.testUtil.ctpUtil.CtpResourcesUtil;
 import io.sphere.sdk.carts.Cart;
 import io.sphere.sdk.carts.CartDraft;
 import io.sphere.sdk.carts.commands.CartCreateCommand;
@@ -30,6 +29,7 @@ import java.util.Optional;
 
 import static com.commercetools.testUtil.CompletionStageUtil.executeBlocking;
 import static com.commercetools.testUtil.ctpUtil.CleanupTableUtil.cleanupOrdersCartsPayments;
+import static com.commercetools.testUtil.ctpUtil.IntegrationCtpResourcesUtil.getCartDraftWithCustomLineItems;
 import static com.commercetools.testUtil.ctpUtil.TaxUtil.TAX_CATEGORY_NAME;
 import static io.sphere.sdk.models.DefaultCurrencyUnits.EUR;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -77,7 +77,7 @@ public class OrderServiceImplIT {
                 TaxCategoryQuery.of().plusPredicates(m -> m.name().is(TAX_CATEGORY_NAME)))
         ).head().get();
 
-        CartDraft cartDraft = CtpResourcesUtil.getCartDraftWithCustomLineItems(taxCategory);
+        CartDraft cartDraft = getCartDraftWithCustomLineItems(taxCategory);
 
         Cart ctCart = executeBlocking(sphereClient.execute(CartCreateCommand.of(cartDraft)));
